KTOR-9784 Support multipart on non-JVM server targets
The default receive transformation delegated multipart handling to a
platform-specific `multiPartData`, and the non-JVM actual threw. As a result,
on Native and JS/Wasm `receiveMultipart()` failed with
CannotTransformContentToTypeException, and `receiveParameters()` with a
multipart body threw IllegalStateException.

The CIO multipart parser was moved to commonMain in KTOR-6632 so that the
client could parse multipart responses on every target, but the server was
never re-wired. Move the JVM `multiPartData` implementation to commonMain and
handle `MultiPartData` in the common transformation. No new dependency is
required and no public API changes.
