+++ ktor-server/ktor-server-core/jvm/src/io/ktor/server/engine/DefaultTransformJvm.kt
@@ -4,19 +4,13 @@
 
 package io.ktor.server.engine
 
-import io.ktor.http.*
-import io.ktor.http.cio.*
-import io.ktor.http.cio.internals.*
-import io.ktor.http.content.*
 import io.ktor.server.application.*
-import io.ktor.server.plugins.UnsupportedMediaTypeException
 import io.ktor.server.request.*
 import io.ktor.util.pipeline.*
 import io.ktor.utils.io.*
 import io.ktor.utils.io.charsets.*
 import io.ktor.utils.io.jvm.javaio.*
 import io.ktor.utils.io.streams.*
-import kotlinx.coroutines.*
 import kotlinx.io.*
 import java.io.*
 
@@ -27,28 +21,7 @@ internal actual suspend fun PipelineContext<Any, PipelineCall>.defaultPlatformTr
 
     return when (call.receiveType.type) {
         InputStream::class -> receiveGuardedInputStream(channel)
-        MultiPartData::class -> multiPartData(channel)
         else -> null
-    }
-}
-
-@OptIn(InternalAPI::class)
-internal actual fun PipelineContext<*, PipelineCall>.multiPartData(rc: ByteReadChannel): MultiPartData {
-    val contentType = call.request.header(HttpHeaders.ContentType)
-        ?: throw UnsupportedMediaTypeException(null)
-
-    val contentLength = call.request.header(HttpHeaders.ContentLength)?.toLong()
-
-    try {
-        return CIOMultipartDataBase(
-            coroutineContext + Dispatchers.Unconfined,
-            rc,
-            contentType,
-            contentLength,
-            formFieldLimit = call.formFieldLimit
-        )
-    } catch (_: UnsupportedMediaTypeExceptionCIO) {
-        throw UnsupportedMediaTypeException(ContentType.parse(contentType))
     }
 }
