+++ ktor-server/ktor-server-core/nonJvm/src/io/ktor/server/engine/DefaultTransform.nonJvm.kt
@@ -4,19 +4,14 @@
 
 package io.ktor.server.engine
 
-import io.ktor.http.content.*
 import io.ktor.server.application.*
 import io.ktor.util.pipeline.*
-import io.ktor.utils.io.*
 import io.ktor.utils.io.charsets.*
 import kotlinx.io.*
 
 internal actual suspend fun PipelineContext<Any, PipelineCall>.defaultPlatformTransformations(
     query: Any
 ): Any? = null
-
-internal actual fun PipelineContext<*, PipelineCall>.multiPartData(rc: ByteReadChannel): MultiPartData =
-    error("Multipart is not supported on non JVM platforms")
 
 internal actual fun Source.readTextWithCustomCharset(charset: Charset): String =
     error("Charset $charset is not supported on non JVM platforms")
